Created API method to declare war, refactored faction dashboard code organization
