+++ src/components/dashboards/factionDashboard/Regions/regionSearchBar.vue
