+++ src/ts/ApiService/FactionApiClient.ts
@@ -2,6 +2,7 @@ import axios from "axios";
 import { useFactionsStore } from "@/stores/generalInfoStores";
 import { Faction } from "@/ts/types/Faction";
 import { ApiClient } from "@/ts/ApiService/ApiClient";
+import { useCharacterStore, usePlayerStore } from "@/stores/playerStores";
 
 export class FactionApiClient extends ApiClient {
   public static async loadFactions(): Promise<Faction[]> {
@@ -46,6 +47,31 @@ export class FactionApiClient extends ApiClient {
             factionsStore.factionNames.push(faction.nameOfFaction);
           });
           resolve(factionsStore.factionNames);
+        });
+    });
+  }
+
+  public static async declareWarToFaction(
+    defendingFactionName: string,
+    nameOfWar: string,
+  ): Promise<void> {
+    return new Promise((resolve, reject) => {
+      // If player does not have enough privileges, do not send the request
+      if (!useCharacterStore().isFactionLeader) {
+        reject("You are not a faction leader");
+      }
+      // Send the request
+      axios
+        .post(this.getBaseUrl() + "/wars/declare", {
+          executorDiscordId: usePlayerStore().discordId,
+          defendingFactionName: defendingFactionName,
+          nameOfWar: nameOfWar,
+        })
+        .then(() => {
+          resolve();
+        })
+        .catch((error) => {
+          reject(error);
         });
     });
   }
